Refactor and move exported function to new file
